[autotest] Ensure we delete iptables rules
If resetting flimflam's list of interfaces to check for captive
portals on throws, we won't delete the iptables rule we installed
at the start of the test.  Juggle things around and add a
try/finally to make sure that we always delete the rules.

BUG=chromium-os:19005
TEST=suite_Smoke

Change-Id: I53602842ede78d3a189077985b2617db854450b8
Reviewed-on: http://gerrit.chromium.org/gerrit/8260
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Peter Mayo <petermayo@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>