autotest: wifi: Use check_frequency to verify roaming
Instead of waiting for a connectivity event from flimflam/supplicant
(which may not happen on some fullmac drivers) just wait around,
ping and then verify the frequency change.

BUG=chromium-os:17516
TEST=Rerun test to completion on ath9k and arm/ath6kl hosts.
WRT failure to merge:  This test (and nothing that depends on
it) is not run in the VM tests, nor in anything but WiFi
enclosure, therefore nothing in commit queue failures have
anything to do with this change.

Change-Id: I1b37b2be3234693c8e8ef5a21f97677b25dbf00d
Reviewed-on: http://gerrit.chromium.org/gerrit/8314
Commit-Ready: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>