Remove the obsolete I915_WORKAROUND from glbench.
The issue this was designed to address no longer exists in the Intel driver, and the workaround is not active in the browser video code.

On Sandy Bridge the workaround provokes a bug in the driver for which a separate Piglit test will be submitted.

Change-Id: I618f76dfeb3a5fcfb87c4446a0e79aaaa1a9666a
Reviewed-on: http://gerrit.chromium.org/gerrit/8329
Tested-by: Stuart Abercrombie <sabercrombie@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Ready: Stuart Abercrombie <sabercrombie@chromium.org>