hardware_Trackpad: support vert/horiz gesture variations in verify_sequence
This CL adds the flexibility in sequence criteria verification for
motion to support vert/horiz subname (variation).

For example,
if the criteria specified in the config for a functionality are
  sequence={...
            ('Motion_x_or_y', '>=', 20),
            ...}
and its subname as
  subname=('vert', ...)

Then, the criteria are interpreted as motion along y axis should
be more than 20.

BUG=chromium-os:21631
TEST=Run trackpad autotest, and it should execute normally.

Change-Id: I31abc453286d579ab6ce22eae308f1489a3344a6
Reviewed-on: http://gerrit.chromium.org/gerrit/10374
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>