hardware_Trackpad: support vert/horiz gesture variations in verify_buttons
This CL adds the flexibility in button criteria verification. It expands
'vert' as both up and down, and 'horiz' as both left and right.

For example,
if the criteria specified in the config for a functionality are
  criteria={'button': ('Button Wheel', '>=', 10),
            ...}
and its subname as
  subname=('vert', ...)

Then, the criteria are expanded as
  Button Wheel Up >= 10, and
  Button Wheel Down >= 10

BUG=chromium-os:21631
TEST=Run trackpad autotest, and it should execute normally.

Change-Id: I5ff7ecd84ba59fe49eb10b484d6554a15cf48ceb
Reviewed-on: http://gerrit.chromium.org/gerrit/10373
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>