wifi autotest: allow test config to specify ssh port
in commit 9e3ff0b81388b0f1fefd4e0f80bdc995f2c0c31d, we switched
from using hosts.create_host, to hosts.SSHHost, when establishing
connections to the router and the server.

in order to support running tests in a VM, this patch adds support
for specifying the port number of the ssh server on the router
and server.

(previously, we piggybacked on the way that create_host
shared connections to the same ip address. due to that, the port
number specified on the run_remote_tests command line was re-used
for the router and server.)

BUG=None
TEST=WiFiManager, on VM and testbed

Change-Id: I2e0ca6e7ee0150f06701b2e4fcabeb8054338c69
Reviewed-on: http://gerrit.chromium.org/gerrit/8444
Commit-Ready: mukesh agrawal <quiche@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>