wifi-roaming: Add test for connect-on-resume
I've experienced issues where the system fails to connect on resume
if it was not connected on suspend.  This may not be a "bug" but a
missing feature, but I think it is a behavior the user expects.  I
have created a roaming suite test to demonstrate the issue.

BUG=chromium-os:21216
TEST=Run test (it fails)

Change-Id: Id1509dcb6dffba14ab3f9c55b927a6164da0accd
Reviewed-on: http://gerrit.chromium.org/gerrit/8820
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>