PlaybackRecordSemiAuto: Handle empty device name.
Change the regex used to extract the name of the device to be
tolerant of cards that don't fill in the device name (The string ends
in '[]').  This gets the device list to work, however the capabilities
are still only queried for HDA cards, I'll open a new bug for that.

BUG=chromium-os:18979
TEST= Run on ZGB to make sure that the change didn't break HDA. Run
on WM8903 based board to make sure the devices are now listed.

Change-Id: I9526543b701c4bbb320130909d8fb0c66656f1a5
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/8917
Reviewed-by: Taylor Hutt <thutt@chromium.org>