Use alsa_caps instead of parsing codec#0 file.
This will be used to list the supported formats, number of channels
and sample rates for alsa devices.  This can be used as a more generic
replacement for parsing /proc/asound/card0/codec#0.
The /proc/asound/card0/codec#0 file is HDA only and doesn't work on
stumpy.  Use the new alsa_caps program to get the sample rates and
formats supported by each card instead.

BUG=chromium-os:21382, chrome-os-partner:6049
TEST=manual, make sure that alsa_cap output provides reasonable info
for alsa devices on ZGB and Kaen.
Run device list test on ZGB, Kaen and Stumpy.

Change-Id: I4f568964b4a0936599a9955597d5636449b2f647
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/10039