autotest (wifi): when running in a VM, use dnsmasq as the DHCP server (rather than ISC dhcpd)
this avoids the need for us to build the net-misc/dhcp package
differently for test and release images.

also: configure hostapd to log to syslog (in addition to its
own log file). this way, we'll have timestampped messages for it.

BUG=chromium-os:21484
TEST=manual (see below)

Ran network_WiFiManager.001_SSID_NonAscii, but with some tweaks
to the VM image. (In particular, installed dnsmasq and shill, and
revved hostapd.)

Change-Id: Ia1a61bd195629228ab990198a7051bc3965db5a0
Reviewed-on: http://gerrit.chromium.org/gerrit/9968
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: mukesh agrawal <quiche@chromium.org>