autotest: wifi: always set RTS and fragmentation threshold
When constructing the hostapd config file always set the RTS and
fragmentation threshold because systems such as linux configure this on
the underlying network device and that persists across runs.  This results
in one test (such as MatFunc/010CheckRxFrag altering the results of all
subsequent tests).

BUG=chromium-os:9424
TEST=downstream udp perf tests and 010CheckRxFrag to verify no regressions

Change-Id: I6e63f7ca38221905e0f6cc911aae057d4cb34602
Reviewed-on: http://gerrit.chromium.org/gerrit/10505
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Sam Leffler <sleffler@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>