Disable SSH host key checking when connecting to FAFT client.
It is for the cases of recovery mode test. The test boots a USB/SD image
instead of an internal image. The RSA host key is changed and causes
failure in SSH connection. So we disable the SSH host key checking by
configuring it to use an empty known_hosts file.

BUG=chromium-os:19710
TEST=run_remote_tests.sh --remote=$REMOTE_IP -a \
     "servo_vid=0x18d1 servo_pid=0x5001" firmware_TryFwB
More test using this config is in a comming CL, firmware_UserRequestRecovery.

Change-Id: I4dff7c151a0395c777451b789f216be084fb819e
Reviewed-on: https://gerrit.chromium.org/gerrit/10632
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>