Support installing Autotest dependency after boot.
It is for the cases of recovery mode test. The test boots a USB/SD image
instead of an internal image. The previous installed Autotest dependency
on the internal image is lost. So we need to install it again.

A configuration of FAFT_STEP, install_deps_after_boot, is supported for
this purpose.

BUG=chromium-os:19710
TEST=run_remote_tests.sh --remote=$REMOTE_IP -a \
     "servo_vid=0x18d1 servo_pid=0x5001" firmware_TryFwB
More test using this config is in a comming CL, firmware_UserRequestRecovery.

Change-Id: I7e3e68c2e15597cdf27979ddbdf5e6e2ed63d745
Reviewed-on: https://gerrit.chromium.org/gerrit/10696
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>