Add delay before connecting to remote RPC on reboot.
It is safer to delay seconds there is a gap between network ready and sshd
ready. There is a possibility that network is OK (ping success) but sshd is
still setting up and causes the later RPC connection failed.

BUG=chromium-os:19710
TEST=run_remote_tests.sh --remote=$REMOTE_IP -a \
     "servo_vid=0x18d1 servo_pid=0x5001" firmware_DevMode
Ran the test for 1 day and no failure.

Change-Id: Ied9f598be2247eab0ab6607232b73bef7f736e9a
Reviewed-on: https://gerrit.chromium.org/gerrit/8118
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>