autotest: wifi: change default snaplen for packet captures
Use 152 for a default snaplen on packet captures instead of 0.  tcpdump
on certain platforms translates 0 to a very large value instead of using
the max frame size causing massive packet loss in captures.  152 is a
compromise: it is large enough to collect large 802.11 frames like ProbeResp
and AssocResp but not so large as to noticeably affect throughput tests
and we're able to collect all of an HT20 throughput test stream on systems
where space is very limited.

While here also change capture start/stop method names to be consistent.

BUG=chromium-os:9424
TEST=run a single ht20 iperf -u stream with run_options=client_capture_all,router_capture_all and verify the capture files have all frames

Change-Id: I906245459756dbdf7988bbb7e0622ef2370030b7
Reviewed-on: https://gerrit.chromium.org/gerrit/10805
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Sam Leffler <sleffler@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>