network_3GSafetyDance: wait longer for assoc
network_3GSafetyDance doesn't specify an assoc_timeout when calling
flimflam.ConnectService, and the default is only 15.  We have a modem
that can take longer than this, especially when just powered on.  So,
bump the timeout up to the same 120 seconds that we allow for config.

BUG=chromium-os:22205
TEST=network_3GSafetyDance doesn't fail with TIMEOUT(assoc)

Change-Id: I36f87fa0fc6a4d65bd77a616fee3902e06db6e03
Reviewed-on: https://gerrit.chromium.org/gerrit/10834
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Commit-Ready: ttuttle <ttuttle@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>