pyauto_test: use stub RecallClientTest as superclass
Instead of deriving directly from test.test, derive from an intermediate
RecallClientTest class and be sure to pass additional arguments to the
initialize() function up to it.

This is implemented as a stub for now, but in future commits will be
expanded to use arguments passed in from the RecallServerTest class to
activate Recall features.

BUG=chromium-os:22310
TEST=verified with tests with and without additional arguments

Change-Id: I2ef739db19998ba4077dff18571f354457ecfc7b
Signed-off-by: Scott James Remnant <keybuk@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/11083
Reviewed-by: Nirnimesh <nirnimesh@chromium.org>