site_test: pass unknown initialize arguments to super
Many tests that subclass UITest and PyAutoTest override initialize(),
and call up to the superclass version as they should, but they don't
allow for additional arguments to be passed to the test which the
superclass may make use of.

RecallClientTest (which is what UITest and PyAutoTest are based on)
expects arguments to be passed in from the RecallServerTest that\
activate it.

These patches allow these tests to work with Recall.

BUG=chromium-os:22310
TEST=picked a couple at random to make sure they still worked

Change-Id: Ib9102756b8fd790763b8beacf91888d63aadd2cc
Signed-off-by: Scott James Remnant <keybuk@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/11085