factory: Add "hwid_cfg" to mark test image.
The same factory test image is used for factory/RMA/... (may be more in future),
so we need some mechanism to differentiate the type of test we performed.

The HWID already has configuration names to describe what we are testing for, so
this CL first tries to look up "cfg" from installed HWID folder, and then decide
which test list to use; and in the end add that to "report_tag".

BUG=chromium-os:22460
TEST=manual

 - create a modified test list in .../suite_Factory/test_list_rma
 - echo "rma" > /usr/local/chromeos-hwid/cfg
 - start factory, seeing test_list_rma applied

Change-Id: Idfdae44a176be7605718bf1f1efac84be684d5ae
Reviewed-on: https://gerrit.chromium.org/gerrit/11122
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Tammo Spalink <tammo@chromium.org>
Commit-Ready: Hung-Te Lin <hungte@chromium.org>