CS421X: update controls and files for latest driver change.
The new pin config in the Cirrus driver removes controls for
unconnected paths.  Mic is now attached to line in.  Mic in is not
attached, DMIC is not attached, neither is SPDIF out. remove these
from asound.state as their controls will no longer be present.
Also update to 3.0 controls.

BUG=chrome-os-partner:6135
TEST=sound_infrastructure, plus manually boot stumpy and check with
arecord and aplay.

Change-Id: Iee96548ce9ddbb388a4e07915626cb42585db03c
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/11194
Reviewed-by: Taylor Hutt <thutt@chromium.org>