power_status : Initialize thermal_path to None
In situations where there are no thermal devices, power_status.py
will never define self.thermal_path, so it will fail in refresh
when it tries to check if self.thermal_path is None. Set to none
so that it passes in this case.

BUG=chromium-os:22563
TEST=run the power_LoadTest on a system with thermal sensors
disabled (for example, seaboard variants with kernel 3.0.3)
Ensure the test runs successfully without failing in init.

Change-Id: I58e4c812573242b41155d93cb17c1289c7135250
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/11254
Reviewed-by: Simon Que <sque@chromium.org>