autotest: add recall client-side support
Implement a context manager that can be used to wrap a test to use a
remote recall server, changing the client's configuration to install
a fake root certificate and point the dns resolver and default route
gateway to the Recall server for the duration of the "with" statement.

Include a test case that exercises the code and makes sure it does
the right thing, including cleaning up; this also helpfully serves
as an example for how you could use Recall inside a test case that
otherwise wouldn't use it.

BUG=chromium-os:22310
TEST=run_remote_tests test_Recall

Change-Id: I1701098340998399075abfaceb82b9a72c8c5499
Reviewed-on: https://gerrit.chromium.org/gerrit/11513
Commit-Ready: Scott James Remnant <keybuk@chromium.org>
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>