[autotest] Ensure we grab crash dumps at end of tests
In the event that one of the client tests falls over hard enough that the
device reboots, ABORTing the test in the middle, we still want to enable
autotest to go get any crash dumps that might have been left.

use the get_site_crashdumps() hook to enable this.

BUG=chromium-os:22694
TEST=while run_remote_tests.sh LoginSuccess against some machine, ssh in and send SIGKILL to chrome and reboot the machine.  The test harness should wait for a while, time out, and then do some work. You should see messages like "23:41:14 INFO | Collecting /var/spool/crash/chrome.20111110.153745.2243.core..."

Change-Id: I9ba4f01dba7fe500e411e8167b8697804f8938b8
Reviewed-on: https://gerrit.chromium.org/gerrit/11514
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>