[autotest] Some calls in frontend.py are not using dargs correctly
frontend.py has a number of classes that inherit from RpcObject.  That
class has a method called run() that takes a method name to call and a
dictionary of args to pass to it.  Some calls to run() are not
providing keywords for args.  This fixes them.

Also, allow poll_job_results to work with varying definitions of success.
Initially, this handy function was hard-coded to allow 0 or 1 tests in a job to
fail while still declaring the job as a whole successful.  Now, it takes a
predicate to describe how many tests are allowed to fail.

BUG=None
TEST=server/frontend_unittest.py

Change-Id: I5230cb9a257101f120383b69b842023885b5a364
Reviewed-on: https://gerrit.chromium.org/gerrit/11669
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>