Support for starting and stopping remote programs
Add support for running a command on a remote machine and then
terminating it.  Autotest supports running remote commands with
'parallel', but it doesn't support cancelling them when we're
done with them

I've written better versions of this in the past (that track PIDs),
but this is enough to get the job done.

Based on code in site_wifitest.py.

TEST=used as a part of cellular_ThroughputController, to be checked in.
BUG=chromium-os:22298

Change-Id: I76680c69e7c8bb527000a78c0b3240f122fd2698
Reviewed-on: https://gerrit.chromium.org/gerrit/11919
Commit-Ready: David Rochberg <rochberg@chromium.org>
Reviewed-by: David Rochberg <rochberg@chromium.org>
Tested-by: David Rochberg <rochberg@chromium.org>