A FAFT test case for forcing normal boot on dev firmware.
This test is only meaningful on Alex/ZGB, which contains two different types of
RW firmware: normal and developer firmware. It requires a USB disk plugged-in,
which contains a Chrome OS test image (built by "build_image --test").
On runtime, this test sets developer firmware in A and then corrupts the
firmware in B. It forces to do a normal boot. Going to recovery is expected.

BUG=chromium-os:19710
TEST=run_remote_tests.sh --remote=$REMOTE_IP -a "xml_config=$OVERLAY_XML \
        servo_vid=0x18d1 servo_pid=0x5001" DevFwNormalBoot

Change-Id: I5be0643c9445aa8e9c5989a99eac21f9dcc11d3f
Reviewed-on: https://gerrit.chromium.org/gerrit/12534
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>