network_3GDisableWhileConnecting: Don't GetStatus on non-Gobis
Only cromo/gobi-cromo-plugin guarantee the invariant that GetStatus will
always succeed, even if the modem is disabling.  Expecting it to succeed,
either once or repeatedly, will not work on a modem-manager modem.

This patch checks if the modem is a Gobi before setting up the get_status
requirement and callback.

BUG=chromium-os:24225
TEST=is one, now passes

Change-Id: I483cd8a2a58a133a6a0a5eee1b62c33e3ffbf185
Reviewed-on: https://gerrit.chromium.org/gerrit/12909
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Commit-Ready: ttuttle <ttuttle@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>