A FAFT test case of corrupting both firmware body A and B.
This test corrupts both firmware body A and B and checks the next boot results.

The expected behavior is different if the firmware preamble USE_RO_NORMAL
flag is enabled. In the case USE_RO_NORMAL ON, the firmware corruption
doesn't hurt the boot results since it boots the RO path directly and does
not load and verify the RW firmware body. In the case USE_RO_NORMAL OFF,
the firmware verification fails on loading RW firmware and enters recovery
mode. In this case, it requires a USB disk plugged-in, which contains a
Chrome OS test image (built by "build_image --test").

BUG=chromium-os:19710;chrome-os-partner:6882
TEST=tested on both USE_RO_NORMAL enabled and disabled devices:
$ run_remote_tests.sh --remote=$REMOTE_IP -a "xml_config=$OVERLAY_XML \
      servo_vid=0x18d1 servo_pid=0x5001" CorruptBothFwBodyAB/control.both

Change-Id: I37661366e41196abacbf78756032ac6af3f1805e
Reviewed-on: https://gerrit.chromium.org/gerrit/12973
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>