[autotest] Be tolerant of hanging symlinks while copying logs.
When copying logs back at the end of a test, sometimes we
run into hanging symlinks.  We should skip these.

It's OK to modify this code such that we don't follow links to do
chmodding, because we just copied over all the files we're operating
on, and are now doing a bottom-up walk of all copied files.  We'll hit
everything we need to hit because of that, and don't need to follow
links to get to files we might otherwise miss.

BUG=None
TEST=./server/autoserv test_suites/control.dummy

Change-Id: I6380732eb7741a3e446d296a84bc1dc741d102b6
Reviewed-on: https://gerrit.chromium.org/gerrit/13137
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>