A FAFT test of booting an invalid USB image.
This test requires a USB disk plugged-in, which contains a Chrome OS test
image (built by "build_image --test"). On runtime, this test corrupts the
USB image and tries to boot into it. A failure is expected. It then
restores the USB image and boots into it again.

BUG=chromium-os:19710
TEST=run_remote_tests.sh --remote=$REMOTE_IP -a "xml_config=$OVERLAY_XML \
         servo_vid=0x18d1 servo_pid=0x5001" InvalidUSB

Change-Id: I661d357275627b4234e5a88f63a7e0bb133e46bf
Reviewed-on: https://gerrit.chromium.org/gerrit/13492
Reviewed-by: Todd Broch <tbroch@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>