Cellular:  Use RF switch to connect DUT to emulator
Also, some refactoring to make this possible:

Change GetDefaultBasestation to StartDefault.  I'm not happy with
this; I think the correct solution may be to make there be one test
environment object that encapsulates the configuration and has a
StartDefault method.

Check eth_test in addition to eth0

TEST=cellular_Smoke, cellular_ThroughputController
BUG=chromium-os:24703

Change-Id: I530570a9e968f195b34fb9ca7ca3bbf9a92c843e
Reviewed-on: https://gerrit.chromium.org/gerrit/13648
Reviewed-by: David Rochberg <rochberg@chromium.org>
Tested-by: David Rochberg <rochberg@chromium.org>
Commit-Ready: Jason Glasgow <jglasgow@chromium.org>