hardware_Trackpad: add IterationLog class in trackpad_util
The open, write, and close operations on result log file are
moved to trackpad_util module so that it is easier to be reused
by other modules. A separate iteration log is generated when
trackpad autotest is conducted on a user gesture set.

BUG=chromium-os:22569
TEST=No need to test this module until the web programs have been checked in
to capture gesture files.

Change-Id: Ie49442e418900097ab562f4293a45717c39e32a5
Reviewed-on: https://gerrit.chromium.org/gerrit/16244
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>