autotest: wifi: fix client_check_frequency
When iw link does not return a frequency (e.g. when disconnected) we
were not raising an exception to fail the step/test.  This had been
done to enable compatibility with old drivers that reported link status
through debugfs.  Fix this by removing the compatiblity code and raising
an exception as the drivers with compatiblity requirements are no longer
relevant and/or should be updated to report status through nl80211.

While here shorten a related log msg.

BUG=none
TEST=run WiFiRoaming/network_000ChannelHop w/ a busted wpa_supplicant

Change-Id: Id975c9eb3baa710d59dcbc2e6a5e6498dc6a944b
Reviewed-on: https://gerrit.chromium.org/gerrit/13586
Commit-Ready: Sam Leffler <sleffler@chromium.org>
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>