Fix install_recovery_image() the time of USB insertion to work on old models.
If an USB presented before the firmware screen, there are different behaviors
in different models:
  - the old models (before Stumpy) need users to remove and insert the USB;
  - the new models (since Stumpy) directly boot to the USB.

The original code fails in the old models since it inserts the USB first
and does not remove it later. So this change is to fix this issue by
inserting the USB after the firmware screen showed. It works on both models.

BUG=chromium-os:22339
TEST=run_remote_tests.sh --remote=$REMOTE_IP -a "xml_config=$OVERLAY_XML \
             servo_vid=0x18d1 servo_pid=0x5001 image=$RECOVERY_IMAGE" \
             platform_InstallRecoveryImage

Change-Id: Ib2832174851a4ad7f4538310bc567e2388ff66ea
Reviewed-on: https://gerrit.chromium.org/gerrit/13864
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>