Install the image to USB and DUT disk when given an image argument in FAFT.
For FAFT test cases, if an image argument is given, it installs the
specified test image to the USB disk and reboot into it via recovery mode.
Then run 'chromeos-install' to install it to DUT disk. Such that both
USB and disk images are updated.

BUG=chromium-os:22339
TEST=run_remote_tests.sh --remote=$REMOTE_IP -a "xml_config=$OVERLAY_XML \
             servo_vid=0x18d1 servo_pid=0x5001 image=$TEST_IMAGE" \
             firmware_RecoveryButton

Change-Id: I0fbb988c14c7c8ab3e3cf0050dd63fbdcd2dc8e9
Reviewed-on: https://gerrit.chromium.org/gerrit/13865
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>