Use power_short_press to boot the DUT instead of power_normal_press.
The power_normal_press delays 1.2 seconds. A fast machine boots to the
firmware screen in 1.2 seconds and it checks that the power button is pressed.
As a result, the machine is then shutdown. So we now change to use
power_short_press which is 0.1 second.

BUG=chromium-os:24754
TEST=run_remote_tests.sh --remote=$REMOTE_IP -a "xml_config=$OVERLAY_XML \
        servo_vid=0x18d1 servo_pid=0x5001" firmware_FwScreenPressPower

Change-Id: I60759d04e9453ebb04173f5933f383e03a6c9dca
Reviewed-on: https://gerrit.chromium.org/gerrit/14049
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>