[autotest] Add DevServer class that uses the new HTTP API
Instead of working from the existing class that SSH's to the dev
server and runs shell commands, create a new class to work against the
new HTTP api for the dev server.  I do this because it's not clear
that we will want to continue using the interface of the existing
class, as it was built to do very different work than what we'd like
to do with the dev server now.

BUG=chromium-os:24985
TEST=dev_server_unittest.py

Change-Id: Ic776d72309bac30eb8362868370fd9f36616c3ba
Reviewed-on: https://gerrit.chromium.org/gerrit/14005
Commit-Ready: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>