Change the firmware screen delay to 2 seconds.
When the DEV_SCREEN_SHORT_DELAY flag on GBB is enabled, the developer screen
will directly boot into the disk image after 2 seconds. So the original
delay 10 seconds is too long. Change it to 2 seconds, the time that the
developer screen is surely showed, and that the timeout is not happened.

BUG=chromium-os:19710
TEST=run_remote_tests.sh --remote=$REMOTE_IP -a "xml_config=$OVERLAY_XML \
        servo_vid=0x18d1 servo_pid=0x5001" firmware_DevMode

Change-Id: I9c0d687e6f7f9e7360e3f646aad5a3abf0608e92
Reviewed-on: https://gerrit.chromium.org/gerrit/14140
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>