autotest (wifi): in SuspendRoam tests, wait for DUT to sleep before shutting down the AP
This change makes the behavior consistent with the comments,
which indicate that we don't shut down the AP until after the DUT
is asleep.

The impetus for these changes is an unexpected failure when testing with
shill. The unexpected part was that the connection manager attempted to
reconnect to the second AP before going to sleep.

BUG=None
TEST=WiFiRoaming (with flimflam)

Change-Id: I6e57036ede63a0a12388df467b3c4a1d4f414450
Reviewed-on: https://gerrit.chromium.org/gerrit/14762
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Commit-Ready: mukesh agrawal <quiche@chromium.org>