autotest: wifi: setup AP's so they have different BSSID
Move the second AP to 5GHz so it has a different BSSID than the AP we
want to "switch back to".  When both AP's have the same BSSID the direct
re-auth logic in wpa_supplicant is confused into thinking it has found
the AP we just turned down; this isn't remedied until we associate at
which point we fail and backoff.  Unfortunately doing all this results
in our test timing out.  Since this (very unliklely) edge case is not
the point of the test change the test so each AP has a different BSSID.

Note using different bands is a hack that depends on our AP having
separate radios for 2.4 and 5GHz bands.

BUG=none
TEST=run 10 iterations of 001SSIDSwitchBack without failure

Change-Id: Id08dec656637db913164e20f9a3d833ebc4229b0
Reviewed-on: https://gerrit.chromium.org/gerrit/14989
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>
Commit-Ready: Sam Leffler <sleffler@chromium.org>