[autotest] Add job_repo_url host attribute
Whenever we update a machine via the dynamic_suite code, tag it
with a host attribute indicating where packages corresponding to
the build we just installed live.  That way, when it's time to
install autotest on that machine, we'll be able to consult the
attribute and install the appropriate source.

BUG=chromium-os:25653
TEST=./server/autoserv test_suites/dev_harness and then look for a wget command in the test log that mentions autotest/packages
TEST=./run_remote_tests.sh --board=x86-generic --ssh_port=9222 --remote=127.0.0.1 --use_emerged suite_Smoke
STATUS=Fixed

Change-Id: I56bde1ea247ac934a4bd800ca5dd355f5eeed946
Reviewed-on: https://gerrit.chromium.org/gerrit/15096
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>