Integrate the factory camera test with OpenCV
The CL removed the old v4l2 interface that is currently being used.
It will now use OpenCV to interact with the camera, which provides
more functionality, better robustness and is easier to maintain.

The CL also adds a FPS feature which may be useful for performance
evaluation. It is currently default to false but could be turned on
when necessary.

BUG=chrome-os-partner:6497
TEST= cros_workon on autotest-factory and emerge it -> works
TEST=run the camera test under the autotest framework -> works

Change-Id: Icf1e69fd04225f1f2d5db92932951eb08dcd3c0c
Reviewed-on: https://gerrit.chromium.org/gerrit/15312
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Tammo Spalink <tammo@chromium.org>
Commit-Ready: Tai-Hsu Lin <sheckylin@chromium.org>
Tested-by: Tai-Hsu Lin <sheckylin@chromium.org>