Tune the delay of entering recovery mode in FAFT.
A previous CL changed the firmware screen delay from 10s to 2s.
https://gerrit.chromium.org/gerrit/14140
But when entering recovery mode, the firmware first reads the kernel in
the low-speed USB disk. As a result, 2s is not enough. This CL tunes the
delay of entering recovery mode.

BUG=chromium-os:19710
TEST=run_remote_tests.sh --remote=$REMOTE_IP -a "xml_config=$OVERLAY_XML \
        servo_serial=$SERIAL" firmware_FwScreenCloseLid
     run_remote_tests.sh --remote=$REMOTE_IP -a "xml_config=$OVERLAY_XML \
        servo_serial=$SERIAL" firmware_FwScreenPressPower

Change-Id: I28dc468aac2989ca8940a6503c58ac5a51558347
Reviewed-on: https://gerrit.chromium.org/gerrit/15372
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>