Have piglit check for a libdir specified by the ebuild
piglit had the directory for libglut.so hardcoded to /usr/lib/
however, this doesn't work on multi-lib systems so this change
checks for an environment variable called GLUT_LIBDIR which is
exported by the ebuild and uses it if it is present.
This should fix builds on 64bit overlays.

BUG=chromium-os:26027
TEST=build autotest-deps-piglit on a 64bit x86 overlay
CQ-DEPEND=Id37ce941c0d422826f9c88c00d3378fba2f06363

Change-Id: Iab211119596159287b29b257b20721ec52ec6ef0
Reviewed-on: https://gerrit.chromium.org/gerrit/15466
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Stuart Abercrombie <sabercrombie@chromium.org>
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>