Modify dynamic_suite changing Suite entry to INFO.
Change the run_and_wait portion of dynamic_suite to no longer log
the start of a as an actual job but just as an INFO so we can pull
the information out of the database if we need to.

Modify wait_for_results to use yield so we get test status updates as
soon as they happen rather than waiting until the whole SUITE finishes.

TEST=utils/unittest_suite.py autotest_lib.server.cros.dynamic_suite_unittest
BUG=chromium-os:26126

Change-Id: I600f4c486656cfaa2c40c024df6f24694c1e6066
Reviewed-on: https://gerrit.chromium.org/gerrit/15714
Tested-by: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>