[autotest] Check for 'image_name' in autoupdate control file
I intended to check to see if a variable called 'image_name' was defined
in the autoupdate control file and switch some behavior based on that check.
Instead, I was checking to see if a variable named after the _contents_ of
the variable image_name was defined in the control file.  Whoops.

BUG=None
TEST=./server/autoserv test_suites/dev_harness and then look for a wget
     command in the test log that mentions autotest/packages
TEST=./run_remote_tests.sh --board=x86-generic --ssh_port=9222 --remote=127.0.0.1 --use_emerged suite_Smoke

Change-Id: I5c034b4d1c54ed273eae62cc3549481d573a0a0e
Reviewed-on: https://gerrit.chromium.org/gerrit/15777
Tested-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>